requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.18
PyJWT==2.8.0
sqlparse==0.4.4
typing-extensions==4.11.0
